Split EnumerableExtensions file into different extensions
