+++ MiKo.Analyzer.Shared/Rules/Documentation/MiKo_2022_CodeFixProvider.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Composition;
 using System.Linq;
