+++ MiKo.Analyzer.Shared/Rules/Documentation/MiKo_2028_ParamPhraseContainsNameOnlyAnalyzer.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
