+++ MiKo.Analyzer.Shared/Rules/Documentation/MiKo_2044_InvalidSeeParameterInXmlAnalyzer.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
