+++ MiKo.Analyzer.Shared/Rules/Documentation/MiKo_2070_ReturnsSummaryAnalyzer.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections;
-using System.Collections.Generic;
 
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Diagnostics;
