+++ MiKo.Analyzer.Shared/Rules/Documentation/MiKo_2211_CodeFixProvider.cs
@@ -1,5 +1,4 @@
 ﻿using System.Composition;
-using System.Linq;
 
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CodeFixes;
