+++ MiKo.Analyzer.Shared/Rules/Documentation/MiKo_2231_InheritdocGetHashCodeAnalyzer.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
