+++ MiKo.Analyzer.Shared/Rules/Documentation/MiKo_2233_CodeFixProvider.cs
@@ -1,5 +1,4 @@
 ﻿using System.Composition;
-using System.Linq;
 
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CodeFixes;
