+++ MiKo.Analyzer.Shared/Rules/Maintainability/MiKo_3037_DoNotUseMagicNumbersForTimeoutsAnalyzer.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
