+++ MiKo.Analyzer.Shared/Rules/Maintainability/MiKo_3062_ExceptionLogMessageEndsWithColonAnalyzer.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Linq;
 
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
