+++ MiKo.Analyzer.Shared/Rules/Maintainability/MiKo_3063_NonExceptionLogMessageEndsWithDotAnalyzer.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Linq;
 
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
