+++ MiKo.Analyzer.Shared/Rules/Maintainability/MiKo_3121_ObjectUnderTestIsNoInterfaceAnalyzer.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.Linq;
 
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
