+++ MiKo.Analyzer.Shared/Rules/Maintainability/MiKo_3201_InvertIfWhenFollowedByFewCodeLinesAnalyzer.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
