+++ MiKo.Analyzer.Shared/Rules/Maintainability/MiKo_3216_AssignedStaticFieldsAreReadOnlyAnalyzer.cs
@@ -1,6 +1,4 @@
-﻿using System.Linq;
-
-using Microsoft.CodeAnalysis;
+﻿using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Diagnostics;
