+++ MiKo.Analyzer.Shared/Rules/Metrics/MiKo_0004_MethodParameterCountAnalyzer.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Diagnostics;
