+++ MiKo.Analyzer.Shared/Rules/Naming/MiKo_1022_ParameterNameLengthAnalyzer.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.Linq;
 
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Diagnostics;
