+++ MiKo.Analyzer.Shared/Rules/Performance/MiKo_5003_LogExceptionAnalyzer.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Linq;
 
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
