+++ MiKo.Analyzer.Shared/Rules/Spacing/MiKo_6007_ObjectUnderTestStatementSurroundedByBlankLinesAnalyzer.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.Linq;
 
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
