+++ MiKo.Analyzer.Tests/Rules/Documentation/MiKo_2020_InheritdocSummaryAnalyzerTests.cs
@@ -2,6 +2,7 @@
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 
+using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CodeFixes;
 using Microsoft.CodeAnalysis.Diagnostics;
